% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Circle Area and Circumference}
\usage{
circle(r)
}
\arguments{
\item{r}{radius of a circle you want}
}
\value{
Area and circumference of certain radii.
}
\description{
Circle Area and Circumference
}
\examples{
x=c(2,3,4)
circle(x)
}
