% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_get_series.R
\name{fred_get_series}
\alias{fred_get_series}
\title{Title:unemployment rate data from FRED}
\usage{
fred_get_series()
}
\value{
return unemployment rate data from 2010-01-01 to current
}
\description{
Title:unemployment rate data from FRED
}
\examples{
fred_get_series()
}
