% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_root_matrix.R
\name{power_root_matrix}
\alias{power_root_matrix}
\title{Title:Power and root matrix}
\usage{
power_root_matrix(x)
}
\arguments{
\item{x}{any numbers}
}
\value{
fifth power to five root matrix of a number
}
\description{
Title:Power and root matrix
}
\examples{
power_root_matrix(64)
}
