% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowbacktest.R
\name{window_backtesting}
\alias{window_backtesting}
\title{Window-backtest}
\usage{
window_backtesting(
  model,
  data,
  orig,
  h,
  xreg = NULL,
  fixed = NULL,
  inc.mean = TRUE,
  reest = 1
)
}
\arguments{
\item{model}{the model being backtested}

\item{data}{time series data}

\item{orig}{orig}

\item{h}{forecast steps}

\item{xreg}{xreg}

\item{fixed}{fixed}

\item{inc.mean}{mean}

\item{reest}{reest}
}
\value{
return RMSE of out-of-sample forecasts and Mean absolute error of out-of-sample forecasts
}
\description{
Window-backtest
}
\examples{
timeseriesdata=datasets::AirPassengers
model=arima(timeseriesdata,order=c(2,1,1))
backtest=window_backtesting(model,timeseriesdata,orig=120,h=12)
}
